refactor: 내비게이션 관련 클래스를 core 모듈로 이동
앱 전반에서 사용되는 내비게이션 관련 클래스 및 인터페이스들을 `presentation/navigation` 패키지에서 `core/navigation` 패키지로 이동하여 모듈 구조를 개선합니다. 이를 통해 내비게이션 관련 코드의 재사용성과 프로젝트 구조의 명확성을 높입니다.

**주요 변경 사항:**

*   **`core/navigation` 패키지 신설:**
    *   `AppRoute`, `NavEvent`, `NavEffect` 인터페이스를 새로 추가하여 내비게이션 구성 요소의 기반을 정의합니다.
    *   `WriteMode` enum 클래스를 `presentation/navigation`에서 `core/navigation`으로 이동시켰습니다.

*   **기존 내비게이션 클래스 수정:**
    *   `MainDestination`, `MainNavEvent`, `MainNavEffect`가 각각 새로 추가된 `AppRoute`, `NavEvent`, `NavEffect` 인터페이스를 구현하도록 수정합니다.
    *   `WriteMode`의 import 경로를 새로운 위치(`com.picday.diary.core.navigation.WriteMode`)로 변경합니다. 이 변경은 `DiaryRoot.kt`, `MainNavHost.kt`, `MainScreen.kt`, `Screen.kt`, `MainNavReducer.kt`, `MainNavReducerTest.kt` 파일에 적용되었습니다.
