+++ app/src/main/java/com/picday/diary/core/navigation/AppRoute.kt
@@ -1,0 +1,3 @@
+package com.picday.diary.core.navigation
+
+interface AppRoute
