+++ app/src/main/java/com/picday/diary/core/navigation/NavEvent.kt
@@ -1,0 +1,3 @@
+package com.picday.diary.core.navigation
+
+interface NavEvent
