+++ app/src/main/java/com/picday/diary/core/navigation/WriteMode.kt
@@ -1,0 +1,6 @@
+package com.picday.diary.core.navigation
+
+enum class WriteMode {
+    VIEW,
+    ADD
+}
