+++ app/src/main/java/com/picday/diary/presentation/main/MainNavReducer.kt
@@ -1,6 +1,6 @@
 package com.picday.diary.presentation.main
 
-import com.picday.diary.presentation.navigation.WriteMode
+import com.picday.diary.core.navigation.WriteMode
 import java.time.LocalDate
 
 sealed interface MainNavEvent {
