--- app/src/main/java/com/picday/diary/presentation/navigation/WriteMode.kt
@@ -1,6 +1,0 @@
-package com.picday.diary.presentation.navigation
-
-enum class WriteMode {
-    VIEW,
-    ADD
-}
